// samples/Billing/web/src/generated/session-client.ts
export const Modules = {
  "Clients": "Clients",
  "Core": "Core",
  "Invoicing": "Invoicing",
} as const
// ...
export const Permissions = {
  "clients": {
    "read": "clients.read",
    "write": "clients.write",
  },
  "invoices": {
    "read": "invoices.read",
    "write": "invoices.write",
  },
} as const